# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen",
        "description": "This is Documentation of Lumen",
        "contact": {
            "email": "brayenprayoga41@gmail.com"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Localhost"
        }
    ],
    "paths": {
...
        "/user/transaksi": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "operationId": "28cf0e2c6e5ff37c572a62cafbc538e6",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/transaksi/simpan": {
            "post": {
                "tags": [
                    "Transaksi"
                ],
                "operationId": "31fc4c99fff6109b0ae80ca9e52e3913",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kode_dealer": {
                                        "title": "Kode Dealer",
                                        "description": "Kode Dealer",
                                        "example": ""
                                    },
                                    "no_antrian": {
                                        "title": "No. Antrian",
                                        "description": "No. Antrian",
                                        "example": ""
                                    },
                                    "status": {
                                        "title": "Status",
                                        "description": "Status",
                                        "example": "WAITING/DONE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/transaksi/edit": {
            "post": {
                "tags": [
                    "Transaksi"
                ],
                "operationId": "c7cd2c2e218b71c69058f1eca185d26c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kode_dealer": {
                                        "title": "Kode Dealer",
                                        "description": "Kode Dealer",
                                        "example": ""
                                    },
                                    "no_antrian": {
                                        "title": "No. Antrian",
                                        "description": "No. Antrian",
                                        "example": ""
                                    },
                                    "status": {
                                        "title": "Status",
                                        "description": "Status",
                                        "example": "WAITING/DONE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/transaksi/hapus": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "operationId": "31ad4a422d917bf9043c900ecb0ed00d",
                "parameters": [
                    {
                        "name": "kode_transaksi",
                        "in": "query",
                        "description": "Kode Transaksi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Hubungi Developer Untuk Akun Login",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    }
}